Change to Double

+++ app/src/main/java/com/example/laterita/settings/SettingsViewModel.kt
@@ -37,8 +37,8 @@ class SettingsViewModel(private val id: Long, private val vehicleDao: VehicleDao
         }
     }
 
-    fun editVehicle(name: String, model: String, vin: String, licence: String, fuel: Int,
-                    reserve: Int) {
+    fun editVehicle(name: String, model: String, vin: String, licence: String, fuel: Double,
+                    reserve: Double) {
         if (settingsAreValid(name, model, vin, licence, fuel, reserve)) {
             _vehicle.value?.make = name
             _vehicle.value?.model = model
@@ -55,7 +55,7 @@ class SettingsViewModel(private val id: Long, private val vehicleDao: VehicleDao
     }
 
     private fun settingsAreValid(name: String, model: String, vin: String, licence: String,
-                                 fuel: Int, reserve: Int): Boolean {
+                                 fuel: Double, reserve: Double): Boolean {
         return name.isNotBlank() && model.isNotBlank() && vin.isNotBlank() && licence.isNotBlank()
                 && reserve <= fuel
     }
